Skip unfinished tests.
